HBX-3068: Refactor the Ant integration tests to factor out common code   - Create new integration test class 'Hbm2JavaTestIT'   - Move all the tests under 'hbm2java' into this new test class   - Move TutorialTestIt one level up
