--- ant/src/it/java/org/hibernate/tool/ant/hbm2java/NoAnnotationsTestIT.java
@@ -1,34 +1,0 @@
-package org.hibernate.tool.ant.hbm2java;
-
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.io.File;
-import java.nio.file.Files;
-
-import org.hibernate.tool.it.ant.TestTemplate;
-import org.junit.jupiter.api.Test;
-
-public class NoAnnotationsTestIT extends TestTemplate {
-	
-    @Test
-    public void testNoAnnotations() throws Exception {
-		setHibernateToolTaskXml(
-				"        <hibernatetool destdir='generated'>                          \n" +
-				"            <jdbcconfiguration propertyfile='hibernate.properties'/> \n" +
-				"            <hbm2java ejb3='false'/>                                 \n" +
-				"        </hibernatetool>                                             \n"
-			);
-		setDatabaseCreationScript(new String[] {
-				"create table PERSON (ID int not null, NAME varchar(20), primary key (ID))"
-		});
-		createProjectAndBuild();
-		assertFolderExists("generated", 1);
-		assertFileExists("generated/Person.java");
-		String generatedPersonJavaFileContents = getFileContents("generated/Person.java");
-		assertFalse(generatedPersonJavaFileContents.contains("import jakarta.persistence.Entity;"));
-		assertTrue(generatedPersonJavaFileContents.contains("public class Person "));
-    }
-
-}
